chore: rip out dead antagonistic-review (Ava+Jon) apparatus; close #4033
The Ava+Jon antagonistic-review graph was dead code. Its only reachable
entry points were `PlanningService` (never instantiated — orphaned when the
`/a2a` endpoint was retired in a5928d3f) and a `/api/flows/antagonistic-review`
route with zero internal callers and no UI client. No sibling repo
(protoAgent, protoWorkstacean, content-creation, contentMachine, ...) calls it
either. The one *live* plan gate — lead-engineer PLAN → `verifyPlan()` — uses
plain `simpleQuery` and never touched the graph, Jon, or `parseReviewXml`.

Removed (dead):
- libs/flows/src/antagonistic-review/ (graph, state, pairs, all nodes incl.
  jon-review, ava-review/parseReviewXml) + its tests
- apps/server/src/services/antagonistic-review-{service,adapter}.ts
- apps/server/src/routes/flows/ (the uncalled route)
- apps/server/src/services/planning-service.ts + plan-store.ts (orphaned)
- libs/types/src/{review,antagonistic-review}.ts (used only by the above)

Refactored:
- The live plan gate (verifyPlan / verifyPlanGoalBackward) extracted into a
  slim PlanReviewService (apps/server/src/services/plan-review-service.ts).
  It implements the existing IPlanReviewService contract; no behavior change.

Rewired:
- services.ts / routes.ts / lead-engineer.module + service + types + processors:
  antagonisticReviewService -> planReviewService, setAntagonisticReviewService
  -> setPlanReviewService. Dropped the /api/flows mount and flows exports from
  @protolabsai/flows; dropped review type exports from @protolabsai/types.

Closes #4033 (obsoleted): caller-trace persistence at project/PRD level
required A2A/protoWorkstacean intake coupling we're declining to add, and half
its sites were the dead code removed here. See the issue's closing comment.

Verified: build:packages OK, typecheck 22/22, package tests 1056, lead-engineer
server tests 42, lint 0 errors.
