+++ .beads/issues.jsonl
@@ -1,7 +1,7 @@
 {"id":"pro-11v","title":"#4035 board 'Archived' view over the archive read endpoints (UI)","status":"open","priority":4,"issue_type":"feature","created_at":"2026-05-30T19:34:32.404450065Z","created_by":"josh","updated_at":"2026-05-30T19:34:32.404450065Z","source_repo":".","compaction_level":0,"original_size":0}
 {"id":"pro-15l","title":"#3985 chore(a2a): retire /a2a endpoint + agent card (confirm dead first)","status":"blocked","priority":3,"issue_type":"task","created_at":"2026-05-30T08:05:55.668808930Z","created_by":"josh","updated_at":"2026-05-30T09:08:21.497155377Z","source_repo":".","compaction_level":0,"original_size":0}
 {"id":"pro-16n","title":"#4006 GOAP routing audit — sweep all skills for ghost-agent targets, test dispatch, log+close","status":"closed","priority":2,"issue_type":"task","created_at":"2026-05-30T08:05:55.307994263Z","created_by":"josh","updated_at":"2026-05-30T08:58:01.491872586Z","closed_at":"2026-05-30T08:58:01.491494947Z","close_reason":"Merged to main this session","source_repo":".","compaction_level":0,"original_size":0,"dependencies":[{"issue_id":"pro-16n","depends_on_id":"pro-k8u","type":"blocks","created_at":"2026-05-30T08:06:16.854206374Z","created_by":"josh","metadata":"{}","thread_id":""}]}
-{"id":"pro-175","title":"#4033 persist callerTraceId at Project/PRD level for research/PRD/review spans","status":"open","priority":2,"issue_type":"feature","created_at":"2026-05-30T18:36:05.019019476Z","created_by":"josh","updated_at":"2026-05-30T19:34:32.206456547Z","source_repo":".","compaction_level":0,"original_size":0}
+{"id":"pro-175","title":"#4033 persist callerTraceId at Project/PRD level for research/PRD/review spans","status":"closed","priority":2,"issue_type":"feature","created_at":"2026-05-30T18:36:05.019019476Z","created_by":"josh","updated_at":"2026-06-02T23:17:46.797214975Z","closed_at":"2026-06-02T23:17:46.796853696Z","close_reason":"Obsoleted — see GH #4033 closing comment. callerTraceId project-level persistence is out of scope (protoMaker is a pure executor; A2A/protoWorkstacean coupling declined). Half the sites were dead code, now removed; the rest have no live callerTraceId source.","source_repo":".","compaction_level":0,"original_size":0}
 {"id":"pro-1cy","title":"#4025 archive-query read surface (GET /api/projects/:slug/archives + board Archived filter)","status":"closed","priority":2,"issue_type":"feature","created_at":"2026-05-30T18:36:05.220869064Z","created_by":"josh","updated_at":"2026-05-30T19:35:20.527677421Z","closed_at":"2026-05-30T19:35:20.527312863Z","close_reason":"#4025 archive read API merged (#4034); was an independent item, the Tier-1 sequencing dep on #4024 is removed","source_repo":".","compaction_level":0,"original_size":0}
 {"id":"pro-1fr","title":"#3815 deploy-workflow script-injection audit + fix across ecosystem repos (zizmor-detectable)","status":"closed","priority":2,"issue_type":"bug","created_at":"2026-05-30T18:36:05.605390597Z","created_by":"josh","updated_at":"2026-05-30T19:07:21.277984877Z","closed_at":"2026-05-30T19:07:21.277623789Z","close_reason":"#3815 protoMaker fixed; fleet linter (release-tools#29) provides ongoing detection; release-tools own findings in #30","source_repo":".","compaction_level":0,"original_size":0,"dependencies":[{"issue_id":"pro-1fr","depends_on_id":"pro-dcj","type":"blocks","created_at":"2026-05-30T18:36:56.973751705Z","created_by":"josh","metadata":"{}","thread_id":""}]}
 {"id":"pro-1h3","title":"#3995 feat(observability): propagate trace.traceId on signal/submit -> Langfuse","status":"closed","priority":2,"issue_type":"feature","created_at":"2026-05-30T08:05:55.427264990Z","created_by":"josh","updated_at":"2026-05-30T08:58:01.659271571Z","closed_at":"2026-05-30T08:58:01.658901416Z","close_reason":"Merged to main this session","source_repo":".","compaction_level":0,"original_size":0}
