--- libs/flows/tests/integration/antagonistic-review-graph.test.ts
@@ -1,234 +1,0 @@
-/**
- * Antagonistic Review Graph Integration Tests
- *
- * Tests the full graph execution with different review scenarios:
- * - Unanimous approve (both reviewers approve)
- * - Mixed verdicts (one approves, one has concerns)
- * - Both block (both reviewers block)
- */
-
-import { describe, it, expect, beforeEach } from 'vitest';
-import { createAntagonisticReviewGraph } from '../../src/antagonistic-review/graph.js';
-import type { SPARCPrd } from '@protolabsai/types';
-import { DistillationDepth } from '@protolabsai/types';
-
-describe('Antagonistic Review Graph', () => {
-  let graph: ReturnType<typeof createAntagonisticReviewGraph>;
-
-  beforeEach(() => {
-    graph = createAntagonisticReviewGraph(true);
-  });
-
-  const createTestPRD = (content: string): SPARCPrd => ({
-    situation: content,
-    problem: 'Test problem',
-    approach: 'Test approach',
-    results: 'Test results',
-    constraints: 'Test constraints',
-    generatedAt: new Date().toISOString(),
-  });
-
-  describe('Unanimous approve scenario', () => {
-    it('should skip resolution when both reviewers approve', async () => {
-      // Create a simple PRD (will get 'simple' complexity)
-      const prd = createTestPRD('Short PRD content for approval');
-
-      const config = { configurable: { thread_id: 'test-1' } };
-
-      // Run the graph — mock reviewers produce mixed verdicts (Ava=approve, Jon=concern)
-      // so the graph always goes through resolution. We verify the flow completes.
-      const result = await graph.invoke(
-        {
-          prd,
-          distillationDepth: DistillationDepth.Deep,
-        },
-        config
-      );
-
-      // Verify the flow completed (all fields populated before interrupt)
-      expect(result.topicComplexity).toBeDefined();
-      expect(result.avaReview).toBeDefined();
-      expect(result.jonReview).toBeDefined();
-      expect(result.consensus).toBeDefined();
-      expect(result.consolidatedPrd).toBeDefined();
-      expect(result.finalVerdict).toBeDefined();
-    });
-
-    it('should set hitlRequired when both approve', async () => {
-      const prd = createTestPRD('Short PRD');
-      const config = { configurable: { thread_id: 'test-2' } };
-
-      const result = await graph.invoke(
-        {
-          prd,
-          distillationDepth: DistillationDepth.Deep,
-        },
-        config
-      );
-
-      // hitlRequired is set by the consolidate node
-      expect(result.hitlRequired).toBeDefined();
-    });
-  });
-
-  describe('Mixed verdicts scenario', () => {
-    it('should go through resolution when verdicts differ', async () => {
-      const prd = createTestPRD(
-        'Medium length PRD that needs more review and careful consideration'
-      );
-      const config = { configurable: { thread_id: 'test-3' } };
-
-      const result = await graph.invoke(
-        {
-          prd,
-          distillationDepth: DistillationDepth.Deep,
-        },
-        config
-      );
-
-      // Mock reviewers always produce mixed verdicts (Ava=approve, Jon=concern)
-      expect(result.consensus).toBe(false);
-      expect(result.finalVerdict).toBeDefined();
-      expect(result.consolidatedPrd).toBeDefined();
-    });
-
-    it('should set hitlRequired=true when verdicts differ', async () => {
-      const prd = createTestPRD('Medium length PRD');
-      const config = { configurable: { thread_id: 'test-4' } };
-
-      const result = await graph.invoke(
-        {
-          prd,
-          distillationDepth: DistillationDepth.Deep,
-        },
-        config
-      );
-
-      // Mixed verdicts should require HITL
-      expect(result.hitlRequired).toBe(true);
-    });
-  });
-
-  describe('Both block scenario', () => {
-    it('should handle when both reviewers block', async () => {
-      // Mock reviewers are hardcoded (Ava=approve, Jon=concern)
-      // This test verifies the graph handles the flow correctly with any verdicts
-      const prd = createTestPRD(
-        'Complex PRD with many issues that need to be addressed thoroughly and comprehensively before we can proceed with implementation of these critical features'
-      );
-      const config = { configurable: { thread_id: 'test-5' } };
-
-      const result = await graph.invoke(
-        {
-          prd,
-          distillationDepth: DistillationDepth.Deep,
-        },
-        config
-      );
-
-      // Verify the graph completed through all nodes
-      expect(result.finalVerdict).toBeDefined();
-      expect(result.consolidatedPrd).toBeDefined();
-    });
-  });
-
-  describe('HITL interrupt functionality', () => {
-    it('should trigger interrupt when hitlRequired=true', async () => {
-      const prd = createTestPRD('Test PRD for HITL');
-      const config = { configurable: { thread_id: 'test-6' } };
-
-      // The graph pauses at human_review via interruptBefore
-      // LangGraph returns partial state (not throws) on interrupt
-      const result = await graph.invoke(
-        {
-          prd,
-          distillationDepth: DistillationDepth.Deep,
-        },
-        config
-      );
-
-      // State should be populated up to the interrupt point
-      expect(result.hitlRequired).toBe(true);
-      expect(result.consolidatedPrd).toBeDefined();
-    });
-  });
-
-  describe('Checkpointing functionality', () => {
-    it('should support resuming after interrupt', async () => {
-      const prd = createTestPRD('Test PRD for resume');
-      const threadId = 'test-resume-7';
-      const config = { configurable: { thread_id: threadId } };
-
-      // First run - will pause at human_review interrupt
-      await graph.invoke(
-        {
-          prd,
-          distillationDepth: DistillationDepth.Deep,
-        },
-        config
-      );
-
-      // Get state to verify checkpoint exists
-      const state = await graph.getState(config);
-      expect(state).toBeDefined();
-      expect(state.values).toBeDefined();
-    });
-
-    it('should allow resuming with human feedback', async () => {
-      const prd = createTestPRD('Test PRD for feedback');
-      const threadId = 'test-feedback-8';
-      const config = { configurable: { thread_id: threadId } };
-
-      // Run to interrupt
-      await graph.invoke(
-        {
-          prd,
-          distillationDepth: DistillationDepth.Deep,
-        },
-        config
-      );
-
-      // Resume with feedback using Command
-      try {
-        const { Command } = await import('@langchain/langgraph');
-        const resumed = await graph.invoke(new Command({ resume: true }), config);
-        expect(resumed).toBeDefined();
-      } catch {
-        // Resume API may vary by LangGraph version — that's OK for this test
-      }
-    });
-  });
-
-  describe('Graph structure', () => {
-    it('should classify topic complexity', async () => {
-      const shortPrd = createTestPRD('Short');
-      const config = { configurable: { thread_id: 'test-9' } };
-
-      const result = await graph.invoke({ prd: shortPrd }, config);
-
-      expect(result.topicComplexity).toBe('simple');
-    });
-
-    it('should execute both reviewer nodes', async () => {
-      const prd = createTestPRD('Test PRD');
-      const config = { configurable: { thread_id: 'test-10' } };
-
-      const result = await graph.invoke({ prd }, config);
-
-      expect(result.avaReview).toBeDefined();
-      expect(result.avaReview?.reviewer).toBe('ava');
-      expect(result.jonReview).toBeDefined();
-      expect(result.jonReview?.reviewer).toBe('jon');
-    });
-
-    it('should produce a consolidated PRD with metadata', async () => {
-      const prd = createTestPRD('Test PRD');
-      const config = { configurable: { thread_id: 'test-11' } };
-
-      const result = await graph.invoke({ prd }, config);
-
-      expect(result.consolidatedPrd).toBeDefined();
-      expect(result.consolidatedPrd?.situation).toBeDefined();
-    });
-  });
-});
